+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [1.1.1] - 2026-09-25
+
 ### Fixed
 
 - **The video player uses your selected browser's cookies, like downloads do** (SNAT-0075)